requirements.txt
altgraph==0.16.1
certifi==2018.8.24
chardet==3.0.4
ecdsa==0.13
esptool==2.5.0
future==0.16.0
idna==2.7
macholib==1.11
netifaces==0.10.7
pefile==2018.8.8
pyaes==1.6.1
https://github.com/pyinstaller/pyinstaller/archive/bbf964c6b89ca33823031fa7ed277c0269192b3e.zip#egg=PyInstaller
PyQt5==5.11.2
PyQt5-sip==4.19.12
pyserial==3.4
requests==2.19.1
urllib3==1.23
zeroconf==0.20.0
